Create new ProjectView class to hold all the top level project windows
The Project class had the EditWindow as a member previously which contained
the TrackView class. The TrackView class needed to access Project signals
which were not properly constructed(because of member order in Project class).
So instead move all the classes that depend on Project into a ProjectView class
and ensure that Project class is properly constructed before creating the
ProjectView class. So the Project class basically becomes a wrapper around
the multi-threaded application logic that emits signals in the GUI thread.

The other option would be to define a ProjectListener interface that contains
all the "signals" and then the Project class can iterate through all the
ProjectListeners.

Create track also adds a label containing a track name in the control list
area now.