Remove CanvasItem event handlers and defer events to the TrackCanvas
This means that a dynamic_cast has to be used to determine the type
of CanvasItem rather than using virtual CanvasItem event handlers
within which the class type would be known. To make use of the
virtual handlers means that unless the events can be handled within
the classes(I don't think they can) then TrackCanvas needs a set
of event handlers for each class type derived from CanvasItem.