Add test for thread safety of glib g_ascii_strtod and g_ascii_dtostr
This is kind of the point of these functions anyway so I'm sure that they have
been tested but it doesn't hurt to confirm.

Boost.Test is not thread-safe apparently. Seems to be fine on linux but not on
Windows with gcc/mingw-w64 so using BOOST_CHECK in two threads at the same time
is not thread-safe. Just use BOOST_CHECK in one thread and assert in the other