Fix core initialize/deinititialize test
The test was basically broken. init/deinit isn't designed to be randomly called
from several threads. A number of threads may call init/deinit in sequence and
this test was meant to test whether or not initialization and deinitialization
only occur once as a result of using atomic variables.

I'm not sure the test is now testing this very well but it does test
initialization/deinitialization in a basic way also.