Add version ~4.0.2 of JUCE to src/third_party/JUCE
It doesn't make much sense to have JUCE as an external module as it is not
designed to be installed as a system wide library. Updates will need to be done
by manually copying from a git checkout. Any changes will need to be noted and
pushed upstream if possible.